Reorganizing the native classes.
