+++ make/modules/jdk.internal.le/Lib.gmk
@@ -27,42 +27,15 @@ include LibCommon.gmk
 
 ################################################################################
 
-ifeq ($(call isTargetOs, windows), true)
+ifeq ($(call isTargetOs, linux macosx windows), true)
 
   $(eval $(call SetupJdkLibrary, BUILD_LIBLE, \
       NAME := le, \
       OPTIMIZATION := LOW, \
       CFLAGS := $(CXXFLAGS_JDKLIB), \
       LDFLAGS := $(LDFLAGS_JDKLIB), \
-      LIBS := $(JDKLIB_LIBS) user32.lib, \
-  ))
-
-  TARGETS += $(BUILD_LIBLE)
-
-endif
-
-ifeq ($(call isTargetOs, linux), true)
-
-  $(eval $(call SetupJdkLibrary, BUILD_LIBLE, \
-      NAME := le, \
-      OPTIMIZATION := LOW, \
-      CFLAGS := $(CXXFLAGS_JDKLIB), \
-      LDFLAGS := $(LDFLAGS_JDKLIB), \
-      LIBS := $(JDKLIB_LIBS) -lstdc++, \
-  ))
-
-  TARGETS += $(BUILD_LIBLE)
-
-endif
-
-ifeq ($(call isTargetOs, macosx), true)
-
-  $(eval $(call SetupJdkLibrary, BUILD_LIBLE, \
-      NAME := le, \
-      OPTIMIZATION := LOW, \
-      CFLAGS := $(CXXFLAGS_JDKLIB), \
-      LDFLAGS := $(LDFLAGS_JDKLIB), \
-      LIBS := $(JDKLIB_LIBS) -lstdc++, \
+      LIBS_windows := $(JDKLIB_LIBS) user32.lib, \
+      LIBS_unix := $(JDKLIB_LIBS) -lstdc++, \
   ))
 
   TARGETS += $(BUILD_LIBLE)
