+++ src/jdk.internal.le/linux/classes/jdk/internal/org/jline/terminal/impl/jna/JDKNativePty.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2018, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2023, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -22,15 +22,30 @@
  * or visit www.oracle.com if you need additional information or have any
  * questions.
  */
-package jdk.internal.org.jline.terminal.impl.jna.osx;
+package jdk.internal.org.jline.terminal.impl.jna;
 
-@SuppressWarnings("serial")
-class LastErrorException extends RuntimeException{
+import java.io.IOException;
+import jdk.internal.org.jline.terminal.Attributes;
+import jdk.internal.org.jline.terminal.Size;
+import jdk.internal.org.jline.terminal.impl.jna.linux.LinuxNativePty;
+import jdk.internal.org.jline.terminal.spi.TerminalProvider;
 
-    public final long lastError;
+class JDKNativePty {
 
-    public LastErrorException(long lastError) {
-        this.lastError = lastError;
+    static JnaNativePty current(TerminalProvider.Stream console) throws IOException {
+        return LinuxNativePty.current(console);
+    }
+
+    static JnaNativePty open(Attributes attr, Size size) throws IOException {
+        return LinuxNativePty.open(attr, size);
+    }
+
+    static int isatty(int fd) {
+        return LinuxNativePty.isatty(fd);
+    }
+
+    static String ttyname(int fd) {
+        return LinuxNativePty.ttyname(fd);
     }
 
 }
