+++ src/jdk.internal.le/linux/classes/jdk/internal/org/jline/terminal/impl/jna/linux/UtilLibraryImpl.java
@@ -24,6 +24,7 @@
  */
 package jdk.internal.org.jline.terminal.impl.jna.linux;
 
+import jdk.internal.org.jline.terminal.impl.jna.LastErrorException;
 import jdk.internal.org.jline.terminal.impl.jna.linux.LinuxNativePty.UtilLibrary;
 
 public final class UtilLibraryImpl implements UtilLibrary {
